# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AIMode": {
        "description": "Which engine actually produced the narrative in a response.\n\nExposed for the same reason as :class:`~app.schemas.common.DataMode`: a consumer\nmust always be able to tell real model output from a canned or fallback response.",
        "enum": [
          "gemini",
          "mock",
          "fallback"
        ],
        "title": "AIMode",
        "type": "string"
      },
...
      "AdvisoryCategory": {
        "enum": [
          "irrigation",
          "disease",
          "pest",
          "nutrient",
          "weather",
          "planting",
          "harvest",
          "soil",
          "regenerative"
        ],
        "title": "AdvisoryCategory",
        "type": "string"
      },
      "AdvisoryList": {
        "description": "Response for `GET /api/v1/farms/{farm_id}/advisories`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Advisory"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdvisoryList",
        "type": "object"
      },
      "AdvisoryPriority": {
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "AdvisoryPriority",
        "type": "string"
      },
      "AnalysisRun": {
        "description": "A complete, immutable farm analysis.",
        "properties": {
          "advisories": {
            "items": {
              "$ref": "#/components/schemas/Advisory"
            },
            "title": "Advisories",
            "type": "array"
          },
          "ai_mode": {
            "$ref": "#/components/schemas/AIMode",
            "description": "Which engine produced the narrative: real model, mock, or deterministic fallback."
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crop_health": {
            "$ref": "#/components/schemas/CropHealth"
          },
          "crop_recommendations": {
            "items": {
              "$ref": "#/components/schemas/CropRecommendation"
            },
            "title": "Crop Recommendations",
            "type": "array"
          },
          "degraded_sources": {
            "description": "Providers that failed or degraded during this run.",
            "items": {
              "type": "string"
            },
            "title": "Degraded Sources",
            "type": "array"
          },
          "disease_risk": {
            "$ref": "#/components/schemas/DiseaseRisk"
          },
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "factors": {
            "description": "Per-factor contributions to the composite, for an explainable UI.",
            "items": {
              "$ref": "#/components/schemas/ScoredFactor"
            },
            "title": "Factors",
            "type": "array"
          },
          "farm_id": {
            "format": "uuid",
            "title": "Farm Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AI model identifier, or null when no model was used.",
            "examples": [
              "gemini-2.5-flash"
            ],
            "title": "Model"
          },
          "overall_band": {
            "$ref": "#/components/schemas/ScoreBand"
          },
          "overall_health_score": {
            "description": "Deterministic weighted composite. Not AI-generated.",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overall Health Score",
            "type": "integer"
          },
          "prompt_version": {
            "description": "Version of the prompt set that produced the narrative fields.",
            "examples": [
              "v1"
            ],
            "title": "Prompt Version",
            "type": "string"
          },
          "regenerative_recommendations": {
            "items": {
              "$ref": "#/components/schemas/RegenerativeRecommendation"
            },
            "title": "Regenerative Recommendations",
            "type": "array"
          },
          "soil_assessment": {
            "$ref": "#/components/schemas/SoilAssessment"
          },
          "sources": {
            "description": "Provenance for every input used.",
            "items": {
              "$ref": "#/components/schemas/DataSourceMeta"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "summary": {
            "description": "Two to three sentence headline for the dashboard.",
            "title": "Summary",
            "type": "string"
          },
          "water_risk": {
            "$ref": "#/components/schemas/WaterRisk"
          },
          "weather_risk": {
            "$ref": "#/components/schemas/WeatherRisk"
          }
        },
        "required": [
          "id",
          "farm_id",
          "status",
          "created_at",
          "duration_ms",
          "prompt_version",
          "ai_mode",
          "overall_health_score",
          "overall_band",
          "summary",
          "weather_risk",
          "water_risk",
          "disease_risk",
          "crop_health",
          "soil_assessment"
        ],
        "title": "AnalysisRun",
        "type": "object"
      },
      "AnalysisRunList": {
        "description": "Response for `GET /api/v1/farms/{farm_id}/analysis`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AnalysisRunList",
        "type": "object"
      },
...
      "AnalysisStatus": {
        "enum": [
          "complete",
          "partial",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "The one and only error shape returned by every 4xx/5xx response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FarmDashboard": {
        "description": "Everything the unified dashboard needs, in one request.\n\nResponse for `GET /api/v1/farms/{farm_id}/dashboard`.\n\nThis endpoint does not 404 when no analysis exists \u2014 it returns\n`has_analysis: false` and `analysis: null` so the frontend renders a\n\"Run analysis\" empty state rather than handling an error.",
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "crops": {
            "items": {
              "$ref": "#/components/schemas/FarmCrop"
            },
            "title": "Crops",
            "type": "array"
          },
          "current_weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherCurrent"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_freshness": {
            "description": "Provenance of every panel, so the UI can badge simulated or stale data.",
            "items": {
              "$ref": "#/components/schemas/DataSourceMeta"
            },
            "title": "Data Freshness",
            "type": "array"
          },
          "farm": {
            "$ref": "#/components/schemas/Farm"
          },
          "has_analysis": {
            "title": "Has Analysis",
            "type": "boolean"
          },
          "recent_images": {
            "items": {
              "$ref": "#/components/schemas/CropImage"
            },
            "title": "Recent Images",
            "type": "array"
          }
        },
        "required": [
          "farm",
          "has_analysis"
        ],
        "title": "FarmDashboard",
        "type": "object"
      },
...
      "ScoreBand": {
        "description": "Qualitative band for a 0-100 score.",
        "enum": [
          "excellent",
          "good",
          "moderate",
          "poor",
          "critical"
        ],
        "title": "ScoreBand",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\n**TerraNex** is the central intelligence layer for agricultural decision support.\n\nIt fuses farm, crop, soil, weather and vegetation data with AI reasoning to produce\nfarm health analysis, weather and water risk, disease risk, agricultural advisories,\ncrop and regenerative-agriculture recommendations, and multimodal crop-image diagnosis.\n\n### For frontend developers\n\n* Every endpoint lives under `/api/v1`. Field casing is `snake_case` throughout.\n* Every failure returns the same envelope:\n  `{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}`.\n  **Branch on `code`** \u2014 `message` is for humans and may change.\n* Every response carries `X-Request-Id`; include it when reporting a bug.\n* Generate typed bindings instead of hand-writing them:\n  `npx openapi-typescript ../contracts/openapi.json -o src/api/types.gen.ts`\n",
    "title": "TerraNex API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis/{run_id}": {
      "get": {
        "description": "Permalink to a specific run. Runs are immutable, so this is stable.",
        "operationId": "get_analysis_run_api_v1_analysis__run_id__get",
        "parameters": [
          {
            "description": "Analysis run identifier.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Analysis run identifier.",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Get an analysis run by id",
        "tags": [
          "analysis",
          "analysis"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/advisories": {
      "get": {
        "description": "Advisories from the latest analysis run, highest priority first. Each carries a `rationale` citing the computed values behind it.",
        "operationId": "list_advisories_api_v1_farms__farm_id__advisories_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AdvisoryCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AdvisoryPriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "include_dismissed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Dismissed",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvisoryList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "AI agricultural advisories",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/analysis": {
      "get": {
        "description": "Lightweight run summaries, newest first \u2014 the source for trend charts.",
        "operationId": "list_analysis_runs_api_v1_farms__farm_id__analysis_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Analysis history",
        "tags": [
          "analysis"
        ]
      },
      "post": {
        "description": "Gathers weather, soil and vegetation data, runs the deterministic risk engine, then asks the AI to interpret the computed results. Persists an immutable run.\n\n**This is the only endpoint that performs farm AI reasoning.** All other analysis endpoints read the stored result.\n\nA run with identical inputs inside the cache window is returned as-is; pass `force_refresh=true` to recompute.\n\nPartial failure does not fail the request: unavailable providers are listed in `degraded_sources` with `status: \"partial\"`, and if the AI is unavailable the narrative falls back to deterministic text with `ai_mode: \"fallback\"`. Check `ai_mode` and `sources[].mode` to see exactly what produced the result.",
        "operationId": "run_analysis_api_v1_farms__farm_id__analysis_post",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "description": "Bypass the cache and recompute from scratch.",
            "in": "query",
            "name": "force_refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Bypass the cache and recompute from scratch.",
              "title": "Force Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "409": {
            "description": "ANALYSIS_IN_PROGRESS"
          },
          "502": {
            "description": "PROVIDER_UNAVAILABLE \u2014 every data source failed"
          },
          "503": {
            "description": "AI_UNAVAILABLE \u2014 AI failed and no fallback was possible"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Run a farm analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/analysis/latest": {
      "get": {
        "description": "The most recent stored run. Returns NO_ANALYSIS_YET if none exists.",
        "operationId": "get_latest_analysis_api_v1_farms__farm_id__analysis_latest_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Latest analysis run",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/dashboard": {
      "get": {
        "description": "Everything the dashboard needs in one request: the farm, its crops, the latest analysis, current weather, recent images, and per-panel data provenance.\n\n**Does not 404 when no analysis exists.** It returns `has_analysis: false` and `analysis: null` so the UI can render a 'Run analysis' empty state instead of handling an error.",
        "operationId": "get_dashboard_api_v1_farms__farm_id__dashboard_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmDashboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Unified farm dashboard",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/risks/disease": {
      "get": {
        "description": "Disease-risk section of the latest analysis run. Each pathogen is evaluated by an explicit agronomic rule over temperature, humidity, duration and growth stage.",
        "operationId": "get_disease_risk_api_v1_farms__farm_id__risks_disease_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseRisk"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Disease risk",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/risks/water": {
      "get": {
        "description": "Water-risk section of the latest analysis run: soil water balance, projected days to stress, and a recommended irrigation depth.",
        "operationId": "get_water_risk_api_v1_farms__farm_id__risks_water_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaterRisk"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Water and irrigation risk",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/risks/weather": {
      "get": {
        "description": "Weather-risk section of the latest analysis run.",
        "operationId": "get_weather_risk_api_v1_farms__farm_id__risks_weather_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherRisk"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Weather risk",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}